ensure attributes are encoded
