+++ cas-server-support-saml/src/main/resources/META-INF/spring/saml-protocol-views.xml
@@ -15,6 +15,8 @@
          is required to be included in the final web application file.  -->
 
     <bean id="abstractsamlView" class="org.jasig.cas.support.saml.web.view.AbstractSaml10ResponseView"
+          p:casAttributeEncoder-ref="casAttributeEncoder"
+          p:servicesManager-ref="servicesManager"
           abstract="true"/>
 
     <bean id="casSamlServiceSuccessView" class="org.jasig.cas.support.saml.web.view.Saml10SuccessResponseView"
