+++ cas-server-webapp-validation/src/main/java/org/jasig/cas/web/view/Cas30ResponseView.java
@@ -1,23 +1,16 @@
 package org.jasig.cas.web.view;
 
-import org.apache.commons.lang3.StringUtils;
 import org.jasig.cas.CasProtocolConstants;
 import org.jasig.cas.authentication.principal.Service;
-import org.jasig.cas.authentication.support.CasAttributeEncoder;
 import org.jasig.cas.services.RegisteredService;
-import org.jasig.cas.services.RegisteredServiceAttributeReleasePolicy;
-import org.jasig.cas.services.ServicesManager;
-import org.jasig.cas.CasViewConstants;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Component;
 import org.springframework.web.servlet.View;
 import org.springframework.web.servlet.view.AbstractUrlBasedView;
 
-import javax.annotation.Resource;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import javax.validation.constraints.NotNull;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
@@ -30,16 +23,6 @@ import java.util.Map;
  * @since 4.1.0
  */
 public class Cas30ResponseView extends Cas20ResponseView {
-
-    /** The attribute encoder instance. */
-    @NotNull
-    @Resource(name="casAttributeEncoder")
-    private CasAttributeEncoder casAttributeEncoder;
-
-    /** The Services manager. */
-    @NotNull
-    @Resource(name="servicesManager")
-    private ServicesManager servicesManager;
 
     /**
      * Instantiates a new Abstract cas response view.
@@ -74,101 +57,6 @@ public class Cas30ResponseView extends Cas20ResponseView {
                 CasProtocolConstants.VALIDATION_CAS_MODEL_ATTRIBUTE_NAME_ATTRIBUTES,
                 this.casAttributeEncoder.encodeAttributes(attributes, getServiceFrom(model)));
     }
-
-    /**
-     * Decide if credential password should be released as attribute.
-     * The credential must have been cached as an authentication attribute
-     * and the attribute release policy must be allowed to release the
-     * attribute.
-     *
-     * @param attributes the attributes
-     * @param model the model
-     * @param service the service
-     */
-    protected void decideIfCredentialPasswordShouldBeReleasedAsAttribute(final Map<String, Object> attributes,
-                                                                         final Map<String, Object> model,
-                                                                         final RegisteredService service) {
-
-        final RegisteredServiceAttributeReleasePolicy policy = service.getAttributeReleasePolicy();
-        final boolean isAuthorized = policy != null && policy.isAuthorizedToReleaseCredentialPassword();
-
-        decideAttributeReleaseBasedOnServiceAttributePolicy(attributes,
-                getAuthenticationAttribute(model, CasViewConstants.MODEL_ATTRIBUTE_NAME_PRINCIPAL_CREDENTIAL),
-                CasViewConstants.MODEL_ATTRIBUTE_NAME_PRINCIPAL_CREDENTIAL,
-                service, isAuthorized);
-    }
-
-    /**
-     * Decide if PGT should be released as attribute.
-     * The PGT must have been cached as an authentication attribute
-     * and the attribute release policy must be allowed to release the
-     * attribute.
-     *
-     * @param attributes the attributes
-     * @param model the model
-     * @param service the service
-     */
-    protected void decideIfProxyGrantingTicketShouldBeReleasedAsAttribute(final Map<String, Object> attributes,
-                                                                         final Map<String, Object> model,
-                                                                         final RegisteredService service) {
-        final RegisteredServiceAttributeReleasePolicy policy = service.getAttributeReleasePolicy();
-        final boolean isAuthorized = policy != null && policy.isAuthorizedToReleaseProxyGrantingTicket();
-
-        decideAttributeReleaseBasedOnServiceAttributePolicy(attributes,
-                getProxyGrantingTicketId(model),
-                CasViewConstants.MODEL_ATTRIBUTE_NAME_PROXY_GRANTING_TICKET,
-                service, isAuthorized);
-    }
-
-    /**
-     * Decide attribute release based on service attribute policy.
-     *
-     * @param attributes the attributes
-     * @param attributeValue the attribute value
-     * @param attributeName the attribute name
-     * @param service the service
-     * @param doesAttributePolicyAllow does attribute policy allow release of this attribute?
-     */
-    protected void decideAttributeReleaseBasedOnServiceAttributePolicy(final Map<String, Object> attributes,
-                                                                       final String attributeValue,
-                                                                       final String attributeName,
-                                                                       final RegisteredService service,
-                                                                       final boolean doesAttributePolicyAllow) {
-        if (StringUtils.isNotBlank(attributeValue)) {
-            logger.debug("Obtained [{}] as an authentication attribute", attributeName);
-
-            if (doesAttributePolicyAllow) {
-                logger.debug("Obtained [{}] is passed to the CAS validation payload", attributeName);
-                attributes.put(attributeName, Collections.singleton(attributeValue));
-            } else {
-                logger.debug("Attribute release policy for [{}] does not authorize the release of [{}]",
-                        service.getServiceId(), attributeName);
-            }
-        } else {
-            logger.trace("[{}] is not available and will not be released to the validation response.", attributeName);
-        }
-    }
-
-    /**
-     * Sets services manager.
-     *
-     * @param servicesManager the services manager
-     * @since 4.1
-     */
-    public void setServicesManager(@NotNull final ServicesManager servicesManager) {
-        this.servicesManager = servicesManager;
-    }
-
-    /**
-     * Sets cas attribute encoder.
-     *
-     * @param casAttributeEncoder the cas attribute encoder
-     * @since 4.1
-     */
-    public void setCasAttributeEncoder(@NotNull final CasAttributeEncoder casAttributeEncoder) {
-        this.casAttributeEncoder = casAttributeEncoder;
-    }
-
 
     /**
      * The type Success.
